Move alluxio config related code to a new place to solve NoClassDefFoundError in MortgageSparkSuite
