+++ sql-plugin/src/main/320+/scala/com/nvidia/spark/rapids/shims/Spark320PlusShims.scala
@@ -385,7 +385,7 @@ trait Spark320PlusShims extends SparkShims with RebaseShims with Logging {
       val options = wrapped.relation.options
 
       val (location, alluxioPathsToReplaceMap) =
-        if (AlluxioUtils.enabledAlluxioReplacementAlgoConvertTime(conf)) {
+        if (AlluxioCfgUtils.enabledAlluxioReplacementAlgoConvertTime(conf)) {
           AlluxioUtils.replacePathIfNeeded(
             conf,
             wrapped.relation,
